Refactor: dumbs index 파일 추가
