+++ src/dumbs/index.tsx
@@ -1,0 +1,17 @@
+import Button from './Button';
+import Layout from './Layout';
+import Logo from './Logo';
+import RequiredInput, { FormLabel, RequiredMark } from './RequiredInput';
+import RequiredSelect from './RequiredSelect';
+import Snackbar from './Snackbar';
+
+export {
+  FormLabel,
+  RequiredMark,
+  Button,
+  Layout,
+  Logo,
+  RequiredInput,
+  RequiredSelect,
+  Snackbar,
+};
